Remove +1 from the string count
withCString adds a null-terminating byte, the +1 includes that byte as part of the string that is stored in the database. It's ignored with a normal string length or select on a row, but you can see it if you select hex(column) or try an equality check on a TEXT value.